requirements.txt
aiodns==2.0.0
aiohttp==3.7.3
aioredis==1.3.1
appdirs==1.4.4
asgiref==3.3.1
async-timeout==3.0.1
asyncio==3.4.3
attrs==19.3.0
autobahn==20.6.1
Automat==20.2.0
binance==0.3
ccxt==1.42.5
certifi==2020.4.5.1
cffi==1.14.0
channels==3.0.4
channels-redis==3.3.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.9
daphne==3.0.2
dateparser==0.7.6
distlib==0.3.1
Django==3.1.7
django-crispy-forms==1.11.0
docopt==0.6.2
filelock==3.0.12
futures3==1.0.0
hiredis==2.0.0
humanize==2.4.1
hyperlink==19.0.0
idna==2.9
importlib-metadata==1.5.0
incremental==17.5.0
msgpack==1.0.2
multidict==4.7.5
mysqlclient==2.0.3
numpy==1.20.1
pandas==1.2.2
Pillow==8.1.0
pipenv==2020.11.15
pipreqs==0.4.10
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycares==3.1.1
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python-binance==0.7.9
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2019.3
regex==2020.6.8
requests==2.23.0
sentry-sdk==1.0.0
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.1
termcolor==1.1.0
tulipy==0.4.0
Twisted==20.3.0
txaio==20.4.1
typing-extensions==3.7.4.3
tzlocal==2.1
ujson==4.0.1
urllib3==1.24.1
virtualenv==20.4.2
virtualenv-clone==0.5.4
whitenoise==5.1.0
yarg==0.1.9
yarl==1.1.0
zipp==3.0.0
zope.interface==5.1.0
